Add files via upload
Done - MESSAGING: ID of match is sent to de show view (-1 if you are trying to see your profile). A set of comments (posts) is created for that match. This comments can only be made and view by the 2 people that matched.
All the comments will have the id of the match. (there are 2 matches between two people so there is a special case in the controller)
Done - Messages opened implementation
Done - Ordered matches by percentage
Done - Notification when you have new messages from a match (text on top of profiles page)
Done - Update location when call reloaded
Done - Update city